requirements.txt
numpy==1.19.5
tensorflow==2.6.0
streamlit==0.89.0
opencv_contrib_python==4.5.3.56
Pillow==8.3.2
